All chests are lockable
